移动部分函数
